requirements.txt
AlbertUnruhUtils.py~=2.1.0
attrs~=21.4.0
pygame~=2.1.2
